pg_hba_resolver: read pg_hba.conf and resolve DNS names to addresses
Read pg_hba.conf and look for comment lines ending with a '# RESOLVE' tag:

Hostnames on such lines are looked up and pg_hba.conf is updated with
uncommented versions of such lines with the hostname replaced by the
resolved address.

Existing entries for the same names are removed if the name no longer
resolves to them.  Postgres is reloaded by sending SIGHUP to the postmaster
if there are any changes.